requirements.txt
Django==2.2.4
django-appconf==1.0.3
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.7.2
django-extensions==2.2.1
django-filter==2.2.0
django-fsm==2.6.1
django-fsm-log==1.6.2
django-tables2==2.1.0
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
